+++ guest/rust/api/src/internal/component/query.rs
@@ -23,8 +23,8 @@ pub fn query<Components: ComponentsTuple + Copy + Clone + 'static>(
 /// change.
 pub fn change_query<Components: ComponentsTuple + Copy + Clone + 'static>(
     components: Components,
-) -> ChangeQuery<Components> {
-    ChangeQuery::create(components)
+) -> UntrackedChangeQuery<Components> {
+    UntrackedChangeQuery::create(components)
 }
 
 /// Creates a new [EventQuery] that will find entities that have the specified `components`
@@ -114,29 +114,50 @@ impl<Components: ComponentsTuple + Copy + Clone + 'static> GeneralQueryBuilder<C
     }
 }
 
-/// An ECS query that calls a callback when entities containing components
-/// marked with [ChangeQuery::track_change] have those components change.
-pub struct ChangeQuery<Components: ComponentsTuple + Copy + Clone + 'static>(
+/// An ECS query that will call a callback when entities containing components
+/// marked with [Self::track_change] have those components change. This type
+/// represents a query that has not had any components marked for tracking yet.
+///
+/// Note that this cannot be built without calling [Self::track_change]
+/// at least once.
+pub struct UntrackedChangeQuery<Components: ComponentsTuple + Copy + Clone + 'static>(
     QueryBuilderImpl<Components>,
     Vec<u32>,
 );
-impl<Components: ComponentsTuple + Copy + Clone + 'static> ChangeQuery<Components> {
-    /// Creates a new [ChangeQuery] that will find entities that have the specified `components`
-    /// that will call its bound function when components marked by [track_change](Self::track_change)
-    /// change.
+impl<Components: ComponentsTuple + Copy + Clone + 'static> UntrackedChangeQuery<Components> {
+    /// Creates a new query that will find entities that have the specified `components`.
+    /// It will call its bound function when components marked by [Self::track_change] change.
     pub fn create(components: Components) -> Self {
         Self(QueryBuilderImpl::new(components.as_indices()), vec![])
     }
 
+    /// The query will return results when these components change values.
+    ///
+    /// This converts this type to a [ChangeQuery], which can be used to bind a callback.
+    pub fn track_change(self, changes: impl ComponentsTuple) -> ChangeQuery<Components> {
+        let cqwt = ChangeQuery(self);
+        cqwt.track_change(changes)
+    }
+}
+
+/// A partially-built ECS query that calls a callback when entities containing components
+/// marked with [Self::track_change] have those components change.
+///
+/// This cannot be constructed without first going through [UntrackedChangeQuery] or
+/// [change_query].
+pub struct ChangeQuery<Components: ComponentsTuple + Copy + Clone + 'static>(
+    UntrackedChangeQuery<Components>,
+);
+impl<Components: ComponentsTuple + Copy + Clone + 'static> ChangeQuery<Components> {
     /// The entities must include the components in `requires`.
     pub fn requires(mut self, requires: impl ComponentsTuple) -> Self {
-        self.0.requires(requires);
+        self.0 .0.requires(requires);
         self
     }
 
     /// The entities must not include the components in `exclude`.
     pub fn excludes(mut self, excludes: impl ComponentsTuple) -> Self {
-        self.0.excludes(excludes);
+        self.0 .0.excludes(excludes);
         self
     }
 
@@ -145,7 +166,7 @@ impl<Components: ComponentsTuple + Copy + Clone + 'static> ChangeQuery<Component
     /// Note that this does *not* implicitly [requires](Self::requires) the components; this allows you to track
     /// changes for entities that do not have all of the tracked components.
     pub fn track_change(mut self, changes: impl ComponentsTuple) -> Self {
-        self.1.extend_from_slice(&changes.as_indices());
+        self.0 .1.extend_from_slice(&changes.as_indices());
         self
     }
 
@@ -165,13 +186,10 @@ impl<Components: ComponentsTuple + Copy + Clone + 'static> ChangeQuery<Component
     }
 
     fn build(self) -> QueryImpl<Components> {
-        assert!(
-            !self.1.is_empty(),
-            "No components specified for tracking. Did you call `ChangeQuery::track_change`?"
-        );
         QueryImpl::new(
             self.0
-                .build_impl(&self.1, wit::component::QueryEvent::Frame),
+                 .0
+                .build_impl(&self.0 .1, wit::component::QueryEvent::Frame),
         )
     }
 }